right-col-content.js also gets event data from vrms-event.jsand is formatted and inserted in right-col-content.js
